Quiet two compiler warnings

+++ server/jsonfmt.c
@@ -490,7 +490,7 @@ int is_json( const char *str )
 
 unsigned int get_js_hash( char const *str )
 {
-  return ((unsigned int)str) % JSON_HASH;
+  return ((uintptr_t)str) % JSON_HASH;
 }
 
 char *prep_for_json_gc( char *str )
